Move CriticalCss to the new Css namespace
